fixing up example case now

+++ utils/dh-key.go
@@ -18,9 +18,9 @@ type Endpoint struct {
 }
 
 // NewBaseModulo is ...
-func NewBaseModulo(bytes uint8) ([]big.Int, error) {
+func NewBaseModulo(bytes int) ([]big.Int, error) {
 	//ideally g**q = 1 mod p, where q is a random prime integer, but all prime numbers will work
-	bits := int(bytes)*8
+	bits := bytes*8
 	if (bytes == 16 || bytes == 24 || bytes == 32){
 		p, _ := rand.Prime(rand.Reader, bits)
 		g, _ := rand.Prime(rand.Reader, bits)
